Ajustando a tela inicial
Search funcionando;
ScrollView com Refresh;
Margin Bottom no FlatList implementado;
