+++ App.js
@@ -1,6 +1,12 @@
 import React from 'react';
+import { StatusBar } from 'react-native';
 import Routes from './src/routes';
 
 export default function App() {
-  return <Routes />;
+  return (
+    <>
+      <StatusBar barStyle="dark-content" />
+      <Routes />
+    </>
+  );
 }
