+++ src/config/ip.js
@@ -1,1 +1,1 @@
-export const ip = '192.168.0.23';
+export const ip = '192.168.1.8';
